feat(*.py): implement interface plot

+++ utils/_test_case_generator.py
@@ -1,12 +1,17 @@
 import json
+import os
 import random
 from datetime import datetime, timedelta
 
 
-def generate_test_case():
+def generate_test_case(ct_name):
+    filename = f'test_cases/{ct_name}.json'
+    if os.path.exists(filename):
+        print('Skiping test case generation!')
+        return None
     # Define the start and end time for the day
     start_time = datetime(2023, 6, 2, 0, 0, 0)
-    end_time = datetime(2023, 6, 2, 23, 59, 59)
+    end_time = datetime(2023, 6, 2, 4, 59, 59)
 
     # Define the time interval range between converters (1 hour to 1 hour 15 minutes)
     min_interval = timedelta(hours=1)
@@ -52,25 +57,25 @@ def generate_test_case():
         converters_2.append(converter)
 
     # Create the JSON object
+    start_time = datetime(2023, 6, 2, 0, 0, 0)
     test_case = {
         "initial_conditions": {
             "time": start_time.isoformat(),
-            "value": 1600
+            "value": 1000
         },
         "converter_1": converters_1,
         "converter_2": converters_2,
-        "pig_iron_hourly_production": 470,
+        "pig_iron_hourly_production": 300,
         "max_restrictive": 2000,
         "allow_auto_spill_events": True,
         "converter_duration": 60
     }
     # Save the data to a JSON file
-    filename = 'test_cases/test_case.json'
+    filename = f'test_cases/{ct_name}.json'
+    if os.path.exists(filename):
+        raise Exception('Test case already exist')
     with open(filename, 'w') as file:
         json.dump(test_case, file, indent=4)
 
     print(f"JSON data has been saved to {filename}.")
 
-
-if __name__ == '__main__':
-    generate_test_case()
